pyproject.toml
[tool.poetry]
name = "artshelter-backend"
version = "0.1.0"
description = ""
authors = ["bokyeal kim <byby8992@gmail.com>"]
readme = "README.md"
packages = [{include = "artshelter_backend"}]

[tool.poetry.dependencies]
python = "^3.8"
djangorestframework = "^3.14.0"
django = "^4.1.7"
pillow = "^9.4.0"
django-environ = "^0.9.0"
django-cors-headers = "^3.13.0"
django-summernote = "^0.8.20.0"
awsebcli = "^3.20.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
